+++ src/main/kotlin/Constants.kt
@@ -1,0 +1,13 @@
+object Constants {
+
+    //Prefix characters for each labels
+    const val LABEL_PREFIX_LINE = '#'
+    const val LABEL_PREFIX_VARIABLE = '@'
+    const val LABEL_PREFIX_LITERAL = '%'
+
+    //Prefix character for frequent variable label
+    const val LABEL_VARIABLE_PREFIX_FREQUENT = '!'
+
+    //Maximum number of label iterations
+    const val MAX_LABEL_ITERATIONS = 100
+}