+++ src/main/kotlin/PreProcessor.kt
@@ -1,25 +1,41 @@
+import Constants.LABEL_PREFIX_LITERAL
+import Constants.LABEL_PREFIX_VARIABLE
+import Constants.LABEL_VARIABLE_PREFIX_FREQUENT
+import Constants.MAX_LABEL_ITERATIONS
+import Tools.findAllValues
+import Tools.processStartingLabel
 import models.Label
 import models.ProcessingFlag
 import models.SourceLine
 import models.VariableType
-import java.io.PrintStream
 
 class PreProcessor(
-    private val optimiser: Optimiser,
+    private val labels: MutableMap<String, Label>,
     private val processingFlags: List<ProcessingFlag>
 ) {
     private val variableNameRepository = VariableNameRepository(
         shortNames = processingFlags.contains(ProcessingFlag.SHORT_VARIABLE_NAMES)
     )
-    private val labels = mutableMapOf<String, Label>()
-    private var basicLineNumber = 0
 
-    fun execute(outputFile: PrintStream, input: List<SourceLine>) {
-        input
+    fun execute(normalSource: List<SourceLine>, frequentSource: List<SourceLine>): List<SourceLine> =
+        if (frequentSource.isEmpty()) {
+
+            //No frequently called sections are specified
+            normalSource
+
+        } else {
+
+            //There are lines marked as frequently called, put this in front of all
+            //other lines and jump over.
+            listOf(
+                SourceLine(content = "goto $FREQUENT_SECTIONS_SKIP_LABEL")
+            ) + frequentSource + listOf(
+                SourceLine(content = FREQUENT_SECTIONS_SKIP_LABEL)
+            ) + normalSource
+        }
             .convertHexadecimalNumbers()
             .processLabels()
-            .outputProcessedSource(outputFile)
-    }
+            .outputProcessedSource()
 
     @OptIn(ExperimentalStdlibApi::class)
     private fun List<SourceLine>.convertHexadecimalNumbers(): List<SourceLine> {
@@ -123,52 +139,58 @@ class PreProcessor(
         return (initFrequentVariablesSrc + this).mapNotNull { line ->
 
             //Line starts with a line number
-            val lineContent = if (line.content[0].isDigit()) {
-                processLineNumber(line)
-            } else {
-                line.content
-            }
+            val lineContent = line.content
 
             //Line starts or continues with a label
             val finalLineContent = if (lineContent.startsWith("{")) {
-                processLabel(line, lineContent)
+                processStartingLabel(labels, line, lineContent) { label, restOfLine ->
+                    if (label[0] == LABEL_PREFIX_LITERAL) {
+
+                        val value = extractLiteralValue(restOfLine)
+
+                        //When value could not be extracted then this is not a literal label definition,
+                        //but rather a label used at the beginning of the line, so skip processing.
+                            ?: return@processStartingLabel null
+
+                        Pair(
+                            Label.Literal(
+                                name = label.drop(1),
+                                originalFormat = "{$label}",
+                                value = value
+                            ),
+                            //The entire line was used by literal label
+                            ""
+                        )
+                    } else {
+                        null
+                    }
+                }
             } else {
                 lineContent
             }
 
-            //When line content is empty after processing line number and label processing
-            //then it can be removed from processing without increasing the BASIC line number.
+            //When line content is empty after label processing then it can be removed.
             if (finalLineContent.isBlank()) {
                 return@mapNotNull null
             }
 
-            //Recreate the line with the new content and BASIC line number and return it.
-            line.copy(
-                basicLineNumber = basicLineNumber,
-                content = finalLineContent
-            ).also {
-                //Increment BASIC line number after line processing.
-                basicLineNumber++
-                if (basicLineNumber > MAX_BASIC_LINE_NUMBER) {
-                    throw Exception("Maximum BASIC line number exceeded, valid range: 0-$MAX_BASIC_LINE_NUMBER")
-                }
-            }
+            //Recreate the line with the new content.
+            line.copy(content = finalLineContent)
         }
     }
 
-    private fun List<SourceLine>.outputProcessedSource(outputFile: PrintStream) {
+    private fun List<SourceLine>.outputProcessedSource(): List<SourceLine> =
+
         //Output pre-processed source with line numbers
-        forEach { line ->
-            //When the line did not receive a BASIC line number then we should skip it
-            val basicLineNumber = line.basicLineNumber ?: return@forEach
+        mapNotNull { line ->
             var lineContent = line.content
 
             checkLabelBlocks(line)
 
             var replaceCount = 0
             do {
-                //Find all labels in the current line
-                val lineLabels = LABEL_REGEX.findAll(lineContent).map { it.groupValues[1] }.toList()
+                //Find all variable and literal labels in the current line
+                val lineLabels = LABEL_REGEX.findAllValues(lineContent)
 
                 //Replace labels in current line with actual line numbers they represent
                 lineLabels.forEach { labelStr ->
@@ -192,13 +214,9 @@ class PreProcessor(
                 //Keep searching for labels in the current line until nothing left
             } while (lineLabels.isNotEmpty())
 
-            //Optimise white space in line when enabled in optimisation flags
-            lineContent = optimiser.optimiseWhiteSpace(lineContent)
-
             //Output processed line
-            outputFile.println("$basicLineNumber $lineContent")
+            line.copy(content = lineContent)
         }
-    }
 
     private fun Variable.createVariableLabel(): Label.Variable {
         //Is this variable name coming with a type postfix?
@@ -249,123 +267,6 @@ class PreProcessor(
         }
     }
 
-    fun printLabels(labelFile: PrintStream) {
-        labelFile.println("Label count: ${labels.size}")
-
-        with(
-            labels.toList().map { it.second }
-        ) {
-            labelFile.println("\nLine labels:")
-
-            filterIsInstance<Label.Line>()
-                .sortedBy { it.basicLineNumber }
-                .forEach { labelFile.println("${it.basicLineNumber}: ${it.name}") }
-
-            labelFile.println("\nVariable labels:")
-
-            filterIsInstance<Label.Variable>()
-                .sortedBy { it.basicName }
-                .forEach {
-                    labelFile.println(
-                        "${it.basicName}: ${it.name}" +
-                                //Add frequent flag when set
-                                if (it.frequent) {
-                                    " - frequent"
-                                } else {
-                                    ""
-                                }
-                    )
-                }
-
-            labelFile.println("\nLiteral labels:")
-
-            filterIsInstance<Label.Literal>()
-                .sortedBy { it.value }
-                .forEach { labelFile.println("${it.value}: ${it.name}") }
-        }
-    }
-
-    private fun processLineNumber(line: SourceLine): String {
-        //Find the end of the line number
-        val index = line.content.indexOfFirst { !it.isDigit() }
-            .takeIf { it >= 0 } ?: line.content.length
-        val lineNumber = line.content.substring(0, index).toInt()
-
-        //Line number cannot be lower than the current BASIC line number
-        if (lineNumber < basicLineNumber) {
-            throw Exception("BASIC line number is lower than current line number: $basicLineNumber\n$line")
-        }
-
-        if (lineNumber > MAX_BASIC_LINE_NUMBER) {
-            throw Exception("Maximum BASIC line number exceeded, valid range: 0-$MAX_BASIC_LINE_NUMBER")
-        }
-
-        //Look for any line labels that were defined with the current BASIC line number
-        labels.entries
-            .filter { (it.value as? Label.Line)?.basicLineNumber == basicLineNumber }
-            .forEach {
-                //Found lines with the old line number, adjust them to the new
-                labels[it.key] = (it.value as Label.Line).copy(basicLineNumber = lineNumber)
-            }
-
-        //Set line number from source as current BASIC line number
-        basicLineNumber = lineNumber
-
-        //Remove line number from line content and return it for further processing
-        return line.content.substring(index).trim()
-    }
-
-    private fun processLabel(line: SourceLine, content: String): String {
-        val index = content.indexOf("}")
-        val label = content.substring(1, index)
-        if (label.isBlank()) throw Exception("Invalid label\n$line")
-
-        val originalLabel = "{$label}"
-        val (labelInstance, restOfLine) = when (label[0]) {
-            LABEL_PREFIX_LINE -> (Label.Line(
-                name = label.drop(1),
-                originalFormat = originalLabel,
-                basicLineNumber = basicLineNumber
-            )
-                    //Return line content without the label
-                    to content.substring(index + 1).trim())
-
-            LABEL_PREFIX_LITERAL -> {
-                val value = extractLiteralValue(content.substring(index + 1).trim())
-
-                //When value could not be extracted then this is not a literal label definition,
-                //but rather a label used at the beginning of the line.
-                    ?: return content
-
-                Pair(
-                    Label.Literal(
-                        name = label.drop(1),
-                        originalFormat = originalLabel,
-                        value = value
-                    ),
-                    //The entire line was used by literal label
-                    ""
-                )
-            }
-
-            //Other kind of labels or escaped characters
-            else -> {
-                //Skip the rest of the label processing and
-                //return line in original form
-                return content
-            }
-        }
-
-        //Each label should be defined only once
-        if (labels.containsKey(originalLabel)) throw Exception("Duplicate label: $originalLabel\n$line")
-
-        //Label must not be empty
-        if (labelInstance.name.isBlank()) throw Exception("Blank label: $label\n$line")
-
-        labels[originalLabel] = labelInstance
-
-        return restOfLine
-    }
 
     private fun extractLiteralValue(content: String): String? =
         when {
@@ -386,16 +287,10 @@ class PreProcessor(
     )
 
     companion object {
-        const val LABEL_PREFIX_LINE = '#'
-        const val LABEL_PREFIX_VARIABLE = '@'
-        const val LABEL_PREFIX_LITERAL = '%'
 
-        const val LABEL_VARIABLE_PREFIX_FREQUENT = '!'
+        private const val FREQUENT_SECTIONS_SKIP_LABEL = "{#preic_skip_frequent_sections}"
 
-        private const val MAX_BASIC_LINE_NUMBER = 63999
-
-        private const val MAX_LABEL_ITERATIONS = 100
-        private val LABEL_PREFIXES = listOf(LABEL_PREFIX_LINE, LABEL_PREFIX_VARIABLE, LABEL_PREFIX_LITERAL)
+        private val LABEL_PREFIXES = listOf(LABEL_PREFIX_VARIABLE, LABEL_PREFIX_LITERAL)
         private val LABEL_PREFIXES_AS_STRING = LABEL_PREFIXES.joinToString("")
 
         private val LABEL_REGEX = Regex("\\{([$LABEL_PREFIXES_AS_STRING].*?)}")